refactor: split index.js into modules
